+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 * !141/!146: Removed all initial implementations tailored to the GFZ-internal preliminary (pre-launch) EnMAP format.
 * !142: Implemented processing of TESTFLAGS-VNIR and TESTFLAGS-SWIR layers in Level-1B reader.
 * !143: Updated copyright.
+* !147: Replaced the user-provided average_elevation by the mean ground elevation provided in the EnMAP metadata.
 
 
 1.2.1 (2026-02-25)